refactor: qr cleanup from mifospay module
