+++ mifospay/src/main/AndroidManifest.xml
@@ -82,7 +82,7 @@
         </provider>
 
         <activity
-            android:name=".qr.ui.ReadQrActivity"
+            android:name=".feature.read.qr.ReadQrActivity"
             android:theme="@style/AppTheme" />
         <activity
             android:name=".feature.passcode.PassCodeActivity"
